+++ src/api/cpp/cvc5.cpp
@@ -38,7 +38,6 @@
 #include "api/cpp/cvc5_checks.h"
 #include "base/check.h"
 #include "base/configuration.h"
-#include "base/modal_exception.h"
 #include "expr/array_store_all.h"
 #include "expr/ascription_type.h"
 #include "expr/cardinality_constraint.h"
@@ -60,11 +59,11 @@
 #include "options/base_options.h"
 #include "options/expr_options.h"
 #include "options/main_options.h"
-#include "options/option_exception.h"
 #include "options/options.h"
 #include "options/options_public.h"
 #include "options/quantifiers_options.h"
 #include "options/smt_options.h"
+#include "proof/proof_node.h"
 #include "proof/unsat_core.h"
 #include "smt/env.h"
 #include "smt/model.h"
@@ -1002,25 +1001,6 @@ class CVC5ApiUnsupportedExceptionStream
  private:
   std::stringstream d_stream;
 };
-
-#define CVC5_API_TRY_CATCH_BEGIN \
-  try                            \
-  {
-#define CVC5_API_TRY_CATCH_END                         \
-  }                                                    \
-  catch (const internal::OptionException& e)           \
-  {                                                    \
-    throw CVC5ApiOptionException(e.getMessage());      \
-  }                                                    \
-  catch (const internal::RecoverableModalException& e) \
-  {                                                    \
-    throw CVC5ApiRecoverableException(e.getMessage()); \
-  }                                                    \
-  catch (const internal::Exception& e)                 \
-  {                                                    \
-    throw CVC5ApiException(e.getMessage());            \
-  }                                                    \
-  catch (const std::invalid_argument& e) { throw CVC5ApiException(e.what()); }
 
 }  // namespace
 
@@ -5037,6 +5017,88 @@ std::ostream& operator<<(std::ostream& out, const Statistics& stats)
   return out;
 }
 
+/*--------------------------------------------------------------------------- */
+/* Proof                                                                      */
+/* -------------------------------------------------------------------------- */
+
+Proof::Proof() {}
+
+Proof::Proof(const std::shared_ptr<internal::ProofNode> p) : d_proof_node(p) {}
+
+Proof::~Proof() {}
+
+ProofRule Proof::getRule() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  if (d_proof_node != nullptr)
+  {
+    return this->getProofNode()->getRule();
+  }
+  return ProofRule::UNKNOWN;
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+Term Proof::getResult() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  if (d_proof_node != nullptr)
+  {
+    internal::NodeManager* nm = internal::NodeManager::currentNM();
+    return Term(nm, this->getProofNode()->getResult());
+  }
+  return Term();
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::vector<Proof> Proof::getChildren() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  if (d_proof_node != nullptr)
+  {
+    std::vector<Proof> children;
+    std::vector<std::shared_ptr<internal::ProofNode>> node_children =
+        d_proof_node->getChildren();
+    for (size_t i = 0, psize = node_children.size(); i < psize; i++)
+    {
+      children.push_back(Proof(node_children[i]));
+    }
+    return children;
+  }
+  return std::vector<Proof>();
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::vector<Term> Proof::getArguments() const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  if (d_proof_node != nullptr)
+  {
+    internal::NodeManager* nm = internal::NodeManager::currentNM();
+    std::vector<Term> args;
+    const std::vector<internal::Node> node_args = d_proof_node->getArguments();
+    for (size_t i = 0, asize = node_args.size(); i < asize; i++)
+    {
+      args.push_back(Term(nm, node_args[i]));
+    }
+    return args;
+  }
+  return std::vector<Term>();
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+const std::shared_ptr<internal::ProofNode>& Proof::getProofNode(void) const
+{
+  return this->d_proof_node;
+}
+
 /* -------------------------------------------------------------------------- */
 /* Solver                                                                     */
 /* -------------------------------------------------------------------------- */
@@ -7124,26 +7186,67 @@ std::pair<Result, std::vector<Term>> Solver::getTimeoutCore() const
       << "Cannot get timeout core unless unsat cores are enabled "
          "(try --produce-unsat-cores)";
   //////// all checks before this line
+  return getTimeoutCoreHelper({});
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+std::pair<Result, std::vector<Term>> Solver::getTimeoutCoreAssuming(
+    const std::vector<Term>& assumptions) const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  CVC5_API_CHECK(!assumptions.empty())
+      << "Cannot get timeout core assuming an empty set of assumptions";
+  CVC5_API_CHECK(d_slv->getOptions().smt.produceUnsatCores)
+      << "Cannot get timeout core unless unsat cores are enabled "
+         "(try --produce-unsat-cores)";
+  //////// all checks before this line
+  return getTimeoutCoreHelper(assumptions);
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+std::pair<Result, std::vector<Term>> Solver::getTimeoutCoreHelper(
+    const std::vector<Term>& assumptions) const
+{
   std::vector<Term> res;
   std::pair<internal::Result, std::vector<internal::Node>> resi =
-      d_slv->getTimeoutCore();
+      d_slv->getTimeoutCore(Term::termVectorToNodes(assumptions));
   for (internal::Node& c : resi.second)
   {
     res.push_back(Term(d_nm, c));
   }
   return std::pair<Result, std::vector<Term>>(Result(resi.first), res);
-  ////////
-  CVC5_API_TRY_CATCH_END;
 }
 
-std::string Solver::getProof(modes::ProofComponent c) const
+std::vector<Proof> Solver::getProof(modes::ProofComponent c) const
 {
   CVC5_API_TRY_CATCH_BEGIN;
   CVC5_API_CHECK(d_slv->getOptions().smt.produceProofs)
       << "Cannot get proof unless proofs are enabled (try --produce-proofs)";
   CVC5_API_RECOVERABLE_CHECK(d_slv->getSmtMode() == internal::SmtMode::UNSAT)
       << "Cannot get proof unless in unsat mode.";
-  return d_slv->getProof(c);
+  //////// all checks before this line
+  std::vector<std::shared_ptr<internal::ProofNode>> proof_nodes =
+      d_slv->getProof(c);
+  std::vector<Proof> proofs;
+  for (auto& p : proof_nodes)
+  {
+    proofs.push_back(Proof(p));
+  }
+  return proofs;
+  ////////
+  CVC5_API_TRY_CATCH_END;
+}
+
+std::string Solver::proofToString(Proof proof, modes::ProofFormat format) const
+{
+  CVC5_API_TRY_CATCH_BEGIN;
+  //////// all checks before this line
+  std::ostringstream ss;
+  this->d_slv->printProof(ss, proof.getProofNode(), format);
+  return ss.str();
+  ////////
   CVC5_API_TRY_CATCH_END;
 }
 
@@ -7625,6 +7728,8 @@ void Solver::setInfo(const std::string& keyword, const std::string& value) const
 void Solver::setLogic(const std::string& logic) const
 {
   CVC5_API_TRY_CATCH_BEGIN;
+  CVC5_API_CHECK(!d_slv->isLogicSet())
+      << "Invalid call to 'setLogic', logic is already set";
   CVC5_API_CHECK(!d_slv->isFullyInited())
       << "Invalid call to 'setLogic', solver is already fully initialized";
   //////// all checks before this line