+++ src/main/portfolio_driver.cpp
@@ -46,7 +46,7 @@ enum SolveStatus : int
 bool ExecutionContext::solveContinuous(parser::InputParser* parser,
                                        bool stopAtSetLogic)
 {
-  std::unique_ptr<Command> cmd;
+  Command cmd;
   bool interrupted = false;
   bool status = true;
   while (status)
@@ -58,15 +58,17 @@ bool ExecutionContext::solveContinuous(parser::InputParser* parser,
       break;
     }
     cmd = parser->nextCommand();
-    if (cmd == nullptr) break;
-
-    status = d_executor->doCommand(cmd);
-    if (cmd->interrupted() && status == 0)
+    if (cmd.isNull())
+    {
+      break;
+    }
+    status = d_executor->doCommand(&cmd);
+    Cmd* cc = cmd.d_cmd.get();
+    if (cc->interrupted() && status == 0)
     {
       interrupted = true;
       break;
     }
-    Cmd* cc = cmd->toCmd();
     if (dynamic_cast<QuitCommand*>(cc) != nullptr)
     {
       break;
@@ -84,16 +86,19 @@ bool ExecutionContext::solveContinuous(parser::InputParser* parser,
   return status;
 }
 
-std::vector<std::unique_ptr<Command>> ExecutionContext::parseCommands(
+std::vector<Command> ExecutionContext::parseCommands(
     parser::InputParser* parser)
 {
-  std::vector<std::unique_ptr<Command>> res;
+  std::vector<Command> res;
   while (true)
   {
-    std::unique_ptr<Command> cmd(parser->nextCommand());
-    if (!cmd) break;
-    res.emplace_back(std::move(cmd));
-    if (dynamic_cast<QuitCommand*>(res.back()->toCmd()) != nullptr)
+    Command cmd = parser->nextCommand();
+    if (cmd.isNull())
+    {
+      break;
+    }
+    res.emplace_back(cmd);
+    if (dynamic_cast<QuitCommand*>(cmd.d_cmd.get()) != nullptr)
     {
       break;
     }
@@ -101,12 +106,11 @@ std::vector<std::unique_ptr<Command>> ExecutionContext::parseCommands(
   return res;
 }
 
-bool ExecutionContext::solveCommands(
-    std::vector<std::unique_ptr<Command>>& cmds)
+bool ExecutionContext::solveCommands(std::vector<Command>& cmds)
 {
   bool interrupted = false;
   bool status = true;
-  for (auto it = cmds.begin(); status && it != cmds.end(); ++it)
+  for (Command& cmd : cmds)
   {
     if (interrupted)
     {
@@ -115,16 +119,15 @@ bool ExecutionContext::solveCommands(
       break;
     }
 
-    Command* cmd = it->get();
-
-    status = d_executor->doCommand(cmd);
-    if (cmd->interrupted() && status == 0)
+    status = d_executor->doCommand(&cmd);
+    Cmd* cc = cmd.d_cmd.get();
+    if (cc->interrupted() && status == 0)
     {
       interrupted = true;
       break;
     }
 
-    if (dynamic_cast<QuitCommand*>(cmd) != nullptr)
+    if (dynamic_cast<QuitCommand*>(cc) != nullptr)
     {
       break;
     }