+++ src/parser/commands.cpp
@@ -1568,7 +1568,39 @@ void GetProofCommand::invoke(cvc5::Solver* solver, SymManager* sm)
 {
   try
   {
-    d_result = solver->getProof(d_component);
+    stringstream ss;
+    const vector<cvc5::Proof> ps = solver->getProof(d_component);
+
+    bool commentProves = !(d_component == modes::ProofComponent::SAT
+                           || d_component == modes::ProofComponent::FULL);
+    modes::ProofFormat format = modes::ProofFormat::DEFAULT;
+    // Ignore proof format, if the proof is not the full proof
+    if (d_component != modes::ProofComponent::FULL)
+    {
+      format = modes::ProofFormat::NONE;
+    }
+
+    if (format == modes::ProofFormat::NONE)
+    {
+      ss << "(" << std::endl;
+    }
+    for (Proof p : ps)
+    {
+      if (commentProves)
+      {
+        ss << "(!" << std::endl;
+      }
+      ss << solver->proofToString(p, format);
+      if (commentProves)
+      {
+        ss << ":proves " << p.getResult() << ")" << std::endl;
+      }
+    }
+    if (format == modes::ProofFormat::NONE)
+    {
+      ss << ")" << std::endl;
+    }
+    d_result = ss.str();
     d_commandStatus = CommandSuccess::instance();
   }
   catch (cvc5::CVC5ApiRecoverableException& e)
@@ -2144,8 +2176,16 @@ void GetDifficultyCommand::toStream(std::ostream& out) const
 /* -------------------------------------------------------------------------- */
 
 GetTimeoutCoreCommand::GetTimeoutCoreCommand()
-    : d_solver(nullptr), d_sm(nullptr)
+    : d_solver(nullptr), d_sm(nullptr), d_assumptions()
 {
+}
+GetTimeoutCoreCommand::GetTimeoutCoreCommand(
+    const std::vector<Term>& assumptions)
+    : d_solver(nullptr), d_sm(nullptr), d_assumptions(assumptions)
+{
+  // providing an empty list of assumptions will make us call getTimeoutCore
+  // below instead of getTimeoutCoreAssuming.
+  Assert(!d_assumptions.empty());
 }
 void GetTimeoutCoreCommand::invoke(cvc5::Solver* solver, SymManager* sm)
 {
@@ -2153,7 +2193,14 @@ void GetTimeoutCoreCommand::invoke(cvc5::Solver* solver, SymManager* sm)
   {
     d_sm = sm;
     d_solver = solver;
-    d_result = solver->getTimeoutCore();
+    if (!d_assumptions.empty())
+    {
+      d_result = solver->getTimeoutCoreAssuming(d_assumptions);
+    }
+    else
+    {
+      d_result = solver->getTimeoutCore();
+    }
     d_commandStatus = CommandSuccess::instance();
   }
   catch (cvc5::CVC5ApiRecoverableException& e)