+++ src/proof/proof_node_to_sexpr.h
@@ -20,10 +20,11 @@
 
 #include <map>
 
+#include "cvc5/cvc5_proof_rule.h"
 #include "expr/kind.h"
 #include "expr/node.h"
 #include "proof/method_id.h"
-#include "cvc5/cvc5_proof_rule.h"
+#include "proof/trust_id.h"
 #include "rewriter/rewrites.h"
 #include "theory/inference_id.h"
 #include "theory/theory_id.h"
@@ -53,7 +54,6 @@ class ProofNodeToSExpr
    */
   Node convertToSExpr(const ProofNode* pn, bool printConclusion = false);
 
- private:
   /** argument format, determines how to print an argument */
   enum class ArgFormat
   {
@@ -65,6 +65,8 @@ class ProofNodeToSExpr
     THEORY_ID,
     // print the argument as a method id
     METHOD_ID,
+    // print the arugment as a trust id
+    TRUST_ID,
     // print the argument as an inference id
     INFERENCE_ID,
     // print the argument as a DSL rewrite id
@@ -72,6 +74,12 @@ class ProofNodeToSExpr
     // print a variable whose name is the term (see getOrMkNodeVariable)
     NODE_VAR
   };
+  /** get argument format for proof node */
+  ArgFormat getArgumentFormat(const ProofNode* pn, size_t i);
+  /** get argument based on the provided format */
+  Node getArgument(Node arg, ArgFormat f);
+
+ private:
   /** map proof rules to a variable */
   std::map<ProofRule, Node> d_pfrMap;
   /** map kind to a variable displaying the kind they represent */
@@ -80,6 +88,8 @@ class ProofNodeToSExpr
   std::map<theory::TheoryId, Node> d_tidMap;
   /** map method ids to a variable displaying the method id they represent */
   std::map<MethodId, Node> d_midMap;
+  /** map trust ids to a variable displaying the method id they represent */
+  std::map<TrustId, Node> d_tridMap;
   /** map infer ids to a variable displaying the inference id they represent */
   std::map<theory::InferenceId, Node> d_iidMap;
   /** map dsl rewrite ids to a variable displaying the dsl rewrite id they
@@ -104,6 +114,8 @@ class ProofNodeToSExpr
   Node getOrMkTheoryIdVariable(TNode n);
   /** get or make method id variable */
   Node getOrMkMethodIdVariable(TNode n);
+  /** get or make trust id variable */
+  Node getOrMkTrustIdVariable(TNode n);
   /** get or make inference id variable */
   Node getOrMkInferenceIdVariable(TNode n);
   /** get or make DSL rewrite id variable */
@@ -114,10 +126,6 @@ class ProofNodeToSExpr
    * print e.g. builtin operators as first-class terms in the SEXPR.
    */
   Node getOrMkNodeVariable(TNode n);
-  /** get argument based on the provided format */
-  Node getArgument(Node arg, ArgFormat f);
-  /** get argument format for proof node */
-  ArgFormat getArgumentFormat(const ProofNode* pn, size_t i);
 };
 
 }  // namespace cvc5::internal