+++ test/unit/api/python/test_uncovered.cpp
@@ -49,6 +49,8 @@ TEST_F(TestApiBlackUncovered, streaming_operators)
   ss << cvc5::modes::LearnedLitType::PREPROCESS;
   ss << cvc5::modes::ProofComponent::FULL;
   ss << cvc5::modes::FindSynthTarget::ENUM;
+  ss << cvc5::modes::InputLanguage::SMT_LIB_2_6;
+  ss << cvc5::modes::ProofFormat::LFSC;
   ss << cvc5::ProofRule::ASSUME;
   ss << cvc5::Result();
   ss << cvc5::Op();
@@ -249,6 +251,15 @@ TEST_F(TestApiBlackUncovered, declareOracleFunUnsat)
       Kind::EQUAL, {d_solver.mkTerm(Kind::APPLY_UF, {f, three}), five});
   d_solver.assertFormula(eq);
   d_solver.checkSat();
+}
+
+TEST_F(TestApiBlackUncovered, Proof)
+{
+  Proof proof;
+  ASSERT_EQ(proof.getRule(), ProofRule::UNKNOWN);
+  ASSERT_TRUE(proof.getResult().isNull());
+  ASSERT_TRUE(proof.getChildren().empty());
+  ASSERT_TRUE(proof.getArguments().empty());
 }
 
 }  // namespace test